separate: ft_cd_chdir func
